db added to server
