+++ database/src/main/java/database/Dao.java
@@ -1,3 +1,5 @@
+package database;
+
 import java.awt.*;
 import java.util.ArrayList;
 import java.util.Optional;
