+++ database/src/main/java/database/HibernateRequest.java
@@ -1,3 +1,5 @@
+package database;
+
 import org.hibernate.Session;
 import org.hibernate.query.Query;
 
@@ -11,12 +13,14 @@ public class HibernateRequest {
 
     public static List<Word> getAllTable() throws SQLException {
         List<Word> words = new ArrayList<>();
+
         try {
-            Query<Word> query = session.createQuery("from Word group by id", Word.class);
+            Query<Word> query = session.createQuery("from database.Word group by id", Word.class);
             words = query.list();
         } catch (Exception e) {
             e.printStackTrace();
         }
+
         HibernateUtil.shutdown();
         return words;
     }
