+++ database/src/main/java/database/HibernateUtil.java
@@ -1,3 +1,4 @@
+package database;
 
 import org.hibernate.SessionFactory;
 import org.hibernate.cfg.Configuration;
@@ -7,10 +8,8 @@ public class HibernateUtil {
 
     private static SessionFactory buildSessionFactory() {
         try {
-            // Create the SessionFactory from hibernate.cfg.xml
             return new Configuration().configure().buildSessionFactory();
         } catch (Throwable ex) {
-            // Make sure you log the exception, as it might be swallowed
             System.err.println("SessionFactory creation failed." + ex);
             throw new ExceptionInInitializerError(ex);
         }
@@ -21,7 +20,6 @@ public class HibernateUtil {
     }
 
     public static void shutdown() {
-        // Close caches and connection pools
         getSessionFactory().close();
     }
