+++ database/src/main/java/database/Word.java
@@ -1,14 +1,14 @@
+package database;
+
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
 
+
 @Entity
 @Table(name = "words")
 public class Word {
-    public Word(){
-
-    }
     @Getter
     @Id
     @GeneratedValue
