+++ database/src/main/java/database/WordDao.java
@@ -1,3 +1,5 @@
+package database;
+
 import java.util.ArrayList;
 import java.util.Objects;
 import java.util.Optional;
