+++ spring_server/pom.xml
@@ -34,10 +34,12 @@
             <version>1.18.18</version>
             <scope>provided</scope>
         </dependency>
+
         <dependency>
             <groupId>org.example</groupId>
-            <artifactId>DB</artifactId>
+            <artifactId>database</artifactId>
             <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
         </dependency>
 
     </dependencies>
