+++ spring_server/src/main/java/spring_server/ServerApp.java
@@ -1,3 +1,5 @@
+package spring_server;
+
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
@@ -5,8 +7,6 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 public class ServerApp {
 
     public static void main(String[] args) {
-        var word = new Word();
-        System.out.println(word.getWord());
         SpringApplication.run(ServerApp.class);
     }
 }
